refactor[]: move LogLevel, LevelToString and getTimestamp to helper as a stand alone function
