+++ source/textFormat.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <formatInterface.h>
 
 namespace sgt {
@@ -5,10 +6,5 @@ namespace sgt {
 	public:
 		virtual ~TextFormat() = default;
 		std::string format(const std::string& text, const LogLevel& level) override;
-
-
-	private:
-		std::string levelToString(const LogLevel& level) const override;
-		std::string getTimeStamp() override;
 	};
 }
